@@ -13,11 +13,10 @@
     "@next/font": "13.1.5",
     "daisyui": "^2.50.0",
     "eslint": "8.32.0",
-    "eslint-config-next": "13.1.5",
     "js-cookie": "^3.0.1",
-    "next": "13.1.5",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "next": "^13.5.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "statsig-react": "^1.27.1"
   },
   "devDependencies": {
@@ -27,6 +26,7 @@
     "@types/node": "18.11.18",
     "@types/react": "18.0.27",
     "autoprefixer": "^10.4.13",
+    "eslint-config-next": "^13.5.4",
     "jest": "^29.4.1",
     "jest-environment-jsdom": "^29.4.1",
     "mockdate": "^3.0.5",
